+++ modules/dynamic_link/src/cpp/RemoveGateway.cpp
@@ -25,51 +25,16 @@
 //=============================================================================
 #include "RemoveGateway.hpp"
 #include "Error.hpp"
-#include "FindDynamicLibraryName.hpp"
-#include "characters_encoding.hpp"
-#include "dynamic_library.hpp"
-#include <boost/filesystem.hpp>
+#include "GatewaysManager.hpp"
 //=============================================================================
 namespace Nelson {
 void
-RemoveGateway(Evaluator* eval, std::wstring dynlibname)
+RemoveGateway(Evaluator* eval, const std::wstring& dynlibname)
 {
-    /* to simplify some dependencies resolution, we move in the directory and restore it after */
-    boost::filesystem::path p;
-    p = dynlibname;
-    p = p.generic_wstring();
-    std::wstring dirname;
-    std::wstring filename;
-    boost::filesystem::path dir = p.parent_path();
-    if (dir.generic_wstring().compare(L"") == 0) {
-        dir = boost::filesystem::current_path();
-    }
-    dirname = dir.generic_wstring();
-    filename = p.filename().generic_wstring();
-    filename = FindDynamicLibraryName(dirname, filename, false);
-    if (filename.empty()) {
-        Error(_W("File not found."));
-    }
-    boost::filesystem::path currentdirbackup = boost::filesystem::current_path();
-    boost::filesystem::current_path(dir);
-    library_handle nlsModuleHandleDynamicLibrary = nullptr;
-#ifdef _MSC_VER
-    nlsModuleHandleDynamicLibrary = load_dynamic_libraryW(filename);
-#else
-    nlsModuleHandleDynamicLibrary = load_dynamic_library(wstring_to_utf8(filename));
-#endif
-    if (nlsModuleHandleDynamicLibrary) {
-        using PROC_RemoveGateway = bool (*)(void*, const wchar_t*);
-        PROC_RemoveGateway RemoveGatewayPtr = reinterpret_cast<PROC_RemoveGateway>(
-            get_function(nlsModuleHandleDynamicLibrary, REMOVEGATEWAY_ENTRY));
-        boost::filesystem::current_path(currentdirbackup);
-        if (!RemoveGatewayPtr) {
-            Error(_W("Module not loaded: symbol not found."));
-        }
-        RemoveGatewayPtr((void*)eval, dynlibname.c_str());
-    } else {
-        boost::filesystem::current_path(currentdirbackup);
-        Error(_W("Module not loaded: library not loaded."));
+    std::wstring errorMessage;
+    GatewaysManager::getInstance()->removeGateway(eval, dynlibname, errorMessage);
+    if (!errorMessage.empty()) {
+        Error(errorMessage);
     }
 }
 //=============================================================================